split off class Operands
